fix: Add generateStaticParams and move some code to UserFunds client component
