+++ frontend/gostarkme-web/app/app/myfunds/[useraddr]/page.tsx
@@ -1,13 +1,14 @@
-"use client";
-
 import React from 'react';
 import UserFunds from '@/components/modules/myfunds/UserFunds';
 import Navbar from '@/components/ui/Navbar';
 import Footer from '@/components/ui/Footer';
-import { Button } from '@/components/ui/Button';
 
 interface MyFundsPageProps {
   params: { useraddr: string };
+}
+
+export function generateStaticParams() {
+  return [{ useraddr: '1' }]
 }
 
 const MyFundsPage: React.FC<MyFundsPageProps> = ({ params }) => {
@@ -17,11 +18,6 @@ const MyFundsPage: React.FC<MyFundsPageProps> = ({ params }) => {
     { label: 'My Profile', href: `/app/myprofile/${useraddr}` },
     { label: 'My funds', href: `/app/myfunds/${useraddr}` }
   ];
-
-  const onNewFundHandler = () => {
-    // TODO: Implement new fund action
-    alert(`Creating new fund`);
-  }
 
   return (
     <div className="min-h-screen flex flex-col">
@@ -36,16 +32,7 @@ const MyFundsPage: React.FC<MyFundsPageProps> = ({ params }) => {
         }}
       />
       <main className="flex flex-grow w-full justify-center bg-white p-8">
-        <div className="bg-white p-4 md:p-6 w-full">
-          <div className="flex justify-between items-center mb-6">
-            <div className="flex items-center">
-              <h1 className="text-2xl font-bold mr-2">My Funds &#10024;</h1>
-            </div>
-            <Button className='!px-12 !py-2.5' label='New' onClick={onNewFundHandler} />
-          </div>
-
-          <UserFunds userAddress={useraddr} />
-        </div>
+        <UserFunds userAddress={useraddr} />
       </main>
       <Footer />
     </div>
